Commit: Ajuste en la entidad

+++ src/main/java/co/edu/ucentral/tarjetac/dto/TarjetasDto.java
@@ -16,7 +16,7 @@ import java.util.Date;
 public class TarjetasDto implements Serializable {
     private long serial;
     @NotBlank(message = "El codigo de la tarjeta es obligatorio")
-    private int numerotarjeta;
+    private long numerotarjeta;
     @NotBlank(message = "La fecha de vencimiento es obligatoria")
     private Date fecha_vec;
     @NotBlank(message = "El documento es obligatorio")
